Create new folder for default operation
This is work towards #816 and #803

We currently have the default mode of analyzing a container
image and a default mode for analyzing Dockerfiles. These modes
are separate from using extensions. So we make a default folder
and move code involved in the default operation there. This commit
is the first pass code move into the default folder.

This will break operation. Do not apply without the rest of the
patchset.
